Day 25
This is terrible, but it's Christmas and I don't want to solve this intense graph theory problem right now. I fed the parsed graph into graphviz so I could see where the connections were, then had the code sever them and calculate the independent graph sizes.

Maybe I'll find a real solve for this later...but probably not.